Sync

+++ load_order_stellaris24.py
@@ -5,6 +5,7 @@ import os
 import ctypes  # An included library with Python install.
 import sys
 import traceback
+import platform
 
 
 def abort(message):
@@ -33,7 +34,13 @@ def getModList(data):
             mod = Mod(key, name, modId)
             modList.append(mod)
         except KeyError:
-            print('key not found in ', key, data)
+            try:
+                name = data['displayName']
+                modId = data['steamId']
+                mod = Mod(key, name, modId)
+                modList.append(mod)
+            except KeyError:
+                print('key not found in ', key, data)
     modList.sort(key=sortedKey, reverse=True)
     return modList
 
@@ -48,13 +55,32 @@ def tweakModOrder(list):
     return list
 
 
-def writeLoadOrder(idList, dlc_load):
+def specialOrder(mods):
+    specialNames = ["UI Overhaul Dynamic", "Dark UI", "Dark U1"]
+    specialList = []
+    for specialName in specialNames:
+        toBeRemoved = []
+        for mod in mods:
+            if specialName in mod.name:
+                specialList.append(mod)
+                toBeRemoved.append(mod)
+
+        for mod in toBeRemoved:
+            mods.remove(mod)
+    return mods + specialList
+
+
+def writeLoadOrder(idList, dlc_load, enabled_mods):
     data = {}
     with open(dlc_load, 'r+') as json_file:
         data = json.load(json_file)
 
     if len(data) < 1:
         abort('dlc_load.json loading failed')
+
+    if enabled_mods is not None:
+        idList = [m for m in idList if m in enabled_mods]
+
     data['enabled_mods'] = idList
 
     with open(dlc_load, 'w') as json_file:
@@ -67,6 +93,7 @@ def writeDisplayOrder(hashList, game_data):
         data = json.load(json_file)
     if len(data) < 1:
         abort('game_data.json loading failed')
+
     data['modsOrder'] = hashList
     with open(game_data, 'w') as json_file:
         json.dump(data, json_file)
@@ -74,8 +101,21 @@ def writeDisplayOrder(hashList, game_data):
 
 def run(settingPath):
     registry = os.path.join(settingPath, 'mods_registry.json')
+
     dlc_load = os.path.join(settingPath, 'dlc_load.json')
-    copyfile(dlc_load, dlc_load + '.bak')
+    if os.path.isfile(dlc_load):
+        copyfile(dlc_load, dlc_load + '.bak')
+    else:
+        abort('please enable at least one mod')
+
+    enabled_mods = None
+    if os.path.exists(dlc_load):
+        with open(dlc_load) as dlc_load_file:
+            dlc_load_data = json.load(dlc_load_file)
+
+            # Do some legwork ahead of time to put into a set to avoidic quadratic loop later for filtering.
+            enabled_mods = frozenset(dlc_load_data.get("enabled_mods", []))
+
     game_data = os.path.join(settingPath, 'game_data.json')
     copyfile(game_data, game_data + '.bak')
 
@@ -83,6 +123,9 @@ def run(settingPath):
     with open(registry, encoding='UTF-8') as json_file:
         data = json.load(json_file)
         modList = getModList(data)
+
+        # move Dark UI and UIOverhual to the bottom
+        modList = specialOrder(modList)
         # make sure UIOverhual+SpeedDial will load after UIOverhual
         modList = tweakModOrder(modList)
     if len(modList) <= 0:
@@ -90,11 +133,14 @@ def run(settingPath):
     idList = [mod.modId for mod in modList]
     hashList = [mod.hashKey for mod in modList]
     writeDisplayOrder(hashList, game_data)
-    writeLoadOrder(idList, dlc_load)
+    writeLoadOrder(idList, dlc_load, enabled_mods)
 
 
 def Mbox(title, text, style):
-    return ctypes.windll.user32.MessageBoxW(0, text, title, style)
+    if platform.system() == 'Windows':
+        return ctypes.windll.user32.MessageBoxW(0, text, title, style)
+    else:
+        print(title + ": " + text)
 
 
 def errorMesssage(error):
